Introducing buffer discard policy option
This change allows the use to control whether the buffer will discard or
retain entries on a failed write attempt.

The handling of 'which' exception causes discard/retention is not dealt
with yet. I.e. This should be extended to allow the user to choose to
discard points that will never succeed (i.e. some 4xx and 5xx response
code) but retain points that failed due to events like network failure.