Creating a narrower API for supported types
The 0.9 Influx docs state that the support types for Fields are:
 - Integers
 - Floats
 - Boolean
 - Strings are text values

These changes deprecate the existing API, that was providing an 'Object'
type interface, and replace it with polymorphic methods that adhere to
the supported types above.

If you are brave, I would even be tempted to remove the @Deprecated
method soon.