Changing name of new methods to 'addField'
This is in accordance with discussion for
https://github.com/influxdb/influxdb-java/pull/116

Doing adding the new methods so they are not the same name as "field()"
means they will not inadvertently change the behaviour of any existing
code.

Included unit test to ensure the behaviour of the deprecated method and
integer input.

Updated all tests to use the new methods (interestingly needed a few
more!), which also made them much clearer regarding the resultant types.