Dependency injected alternative
Chanciness the way to construct an InfluxDB connection to allow the
settings on the underlying HTTP client to be controlled from the
outside.

This is necessary, as there are many things one might want to tweak or
change on the underlying client, but the only way to do them at the
moment is to continue to widen the interface. As is demonstrated here:
https://github.com/influxdb/influxdb-java/pull/123

This set of changes completely removes the changes in #123 and replaces
it with the ability for the programmer to hand the factory methods their
own Retrofit HTTP client instance, which they can pre-configure how they
like.