Fix for #110
This is an alternative fix for Issue #110. This set of changes moves the
validation of the input arguments to the point where they are added
(i.e. it moves the locality of failure closer to the point of the
failure). This is offered as an alterntive implementation to the changes
in PR 133.