First simple implementation of "prepared statements"
I added a new query class BoundParameterQuery, which can be used for
"prepared statements". The constructor accepts a InfluxQL expression
with placeholders, a DB name and a varags list to bind the parameters to
the placeholders. I also extended the InfluxDBService, so that the HTTP
requests contain the "params" parameter. The InfluxDBImpl now
differentiates between a Query and a BoundParameterQuery. This is not
the cleanest solution, because this has to be done at a few locations,
but I didn't want to change too much of the code.